+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.0] - 2020-03-23
+### Added
+- Node `openface_separate` that publishes OpenFace output under separate topics.
+### Fixed
+- Angle conversions for head pose and gaze.
+
 ## [0.2.0] - 2020-03-21
 ### Added
 - Separation of landmark types, i.e. 2D, 2D visible and 3D, with parameters.